Move the syntax generation into a separate class
