+++ lib/src/intl_suggestors/intl_messages.dart
@@ -3,13 +3,9 @@ import 'package:analyzer/dart/analysis/utilities.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:file/file.dart';
 import 'package:file/local.dart';
+import 'package:over_react_codemod/src/intl_suggestors/intl_message_syntax.dart';
 import 'package:over_react_codemod/src/intl_suggestors/utils.dart';
 import 'package:path/path.dart' as p;
-
-// ####### Todo
-// - Get rid of the existing counter
-// - clean up the method names to remove intl prefixes
-// ??
 
 /// The generated messages for the Intl class, with the ability to read and
 /// write the source file for that class.
@@ -23,10 +19,10 @@ class IntlMessages {
   /// The methods for this class, indexed by method name.
   Map<String, Method> methods = {};
 
+  late MessageSyntax syntax;
+
   /// Flag to check if we've actually added anything, and need to rewrite the file.
   bool addedNewMethods = false;
-
-  static const intlFunctionPrefix = 'static String';
 
   // TODO: I think packagePath only applies if there's a sub-package.
   IntlMessages(this.packageName,
@@ -38,6 +34,7 @@ class IntlMessages {
                 'src',
                 'intl',
                 '${packageName}_intl.dart')) {
+    syntax = MessageSyntax(this);
     _readExisting();
   }
 
@@ -68,7 +65,7 @@ class IntlMessages {
     // method, record that.
     addedNewMethods |= true;
     var parsed = MessageParser.forMethod(source).methods.first;
-    var expectedName = giveMeANameForString(parsed.name, parsed.messageText);
+    var expectedName = nameForString(parsed.name, parsed.messageText);
     if (expectedName != parsed.name) {
       throw AssertionError('''
 Attempting to add a different message with the same name:
@@ -79,140 +76,9 @@ Attempting to add a different message with the same name:
     methods.putIfAbsent(parsed.name, () => parsed);
   }
 
-  /// Returns Intl.message for string literal.
-  ///
-  /// The optional [name] parameter lets us provide a name rather than generating
-  /// one from the string.
-  ///
-  /// ex: static String get fooBar => Intl.message('Foo Bar','name: FooBarIntl_fooBar',);
-  String intlGetterDef(StringLiteral node, String namespace, {String? name}) {
-    String text = stringContent(node)!;
-    final varName = giveMeANameFor(node, initialName: name);
-    final message = intlFunctionBody(text, '${namespace}_$varName',
-        isMultiline: isMultiline(node));
-    return '  $intlFunctionPrefix get $varName => $message;';
-  }
-
-  /// A template to build property access for intl string
-  /// ex: ExampleIntl.exampleString
-  String intlStringAccess(StringLiteral node, String namespace,
-          {String? name}) =>
-      '${namespace}.${name ?? giveMeANameFor(node, initialName: name)}';
-
-  /// A template to build function call intl interpolated string
-  /// ex: ExampleIntl.exampleString(sting1, string2)
-  String intlFunctionCall(
-    StringInterpolation node,
-    String namespace,
-    String namePrefix,
-  ) {
-    final baseName = intlFunctionName(node, namePrefix);
-    // TODO: This is very messy to accomodate names starting at intlFunction0
-    // for backward compatibilitys.
-    final functionName = giveMeANameFor(node,
-        initialName: baseName, startAtZero: baseName.endsWith('_intlFunction'));
-    final functionArgs = intlFunctionArguments(node);
-    return '$namespace.$functionName$functionArgs';
-  }
-
-  /// A template to build a function name for Intl message
-  /// ex: Foo_bar
-  String intlFunctionName(
-    StringInterpolation node,
-    String namePrefix,
-  ) =>
-      '${getTestId(null, node) ?? '${namePrefix}_intlFunction'}';
-
-  /// Returns Intl.message with interpolation
-  /// ex: static String Foo_bar(String baz) => Intl.message(
-  ///                                           'Foo bar $baz',
-  ///                                           args: [baz],
-  ///                                           'name: FooBarIntl_Foo_bar',
-  ///                                           );
-  String intlFunctionDef(
-      StringInterpolation node, String namespace, String namePrefix) {
-    final baseName = intlFunctionName(node, namePrefix);
-    // TODO: This is very messy to accomodate function names starting at
-    // intlFunction0 for backward-compatibility.
-    final functionName = giveMeANameFor(node,
-        initialName: baseName, startAtZero: baseName.endsWith('_intlFunction'));
-    final functionParams = intlFunctionParameters(node);
-    final parameterizedMessage = intlParameterizedMessage(node);
-    final messageArgs = intlMessageArgs(node);
-    final message = intlFunctionBody(
-        parameterizedMessage, '${namespace}_$functionName',
-        args: messageArgs, isMultiline: node.isMultiline);
-    return '  $intlFunctionPrefix $functionName$functionParams => $message;';
-  }
-
-  String intlFunctionParameters(StringInterpolation node) {
-    final args = node.elements
-        .whereType<InterpolationExpression>()
-        .map((e) => toVariableName(toNestedName(e.toString())))
-        .toSet()
-        .toList();
-    return '(${args.map((arg) => 'String $arg').toList().join(', ')})';
-  }
-
-  /// A template to build Intl.message('[message]', args: [args], name: [name]);
-  String intlFunctionBody(
-    String message,
-    String name, {
-    List<String> args = const [],
-    bool isMultiline = false,
-  }) {
-    var escapedStr = escapeApos(message);
-    String delimiter = (isMultiline) ? "'''" : "'";
-    return "Intl.message(${delimiter}${escapedStr}${delimiter}, ${args.isNotEmpty ? 'args: ${args}, ' : ''}name: '$name')";
-  }
-
-// --- Start Intl.message parts
-  /// Converts 'Interpolated ${foo.bar} and $baz' into
-  /// 'Interpolated $bar and $baz'
-  String intlParameterizedMessage(StringInterpolation node) => node.elements
-      .map((e) => e is InterpolationExpression
-          ? intlInterpolation(e)
-          : (e as InterpolationString).value)
-      .toList()
-      .join('');
-
-  String intlInterpolation(InterpolationExpression e) {
-    var name = toVariableName(toNestedName('$e'));
-    return r'${' + name + '}';
-  }
-
-  /// Creates the arg array for Intl.message
-  /// ex: For the parameterized string 'Interpolated $bar and $baz'
-  ///     it will return [bar, baz]
-  List<String> intlMessageArgs(StringInterpolation node) => node.elements
-      .whereType<InterpolationExpression>()
-      .map((e) => toVariableName(toNestedName(e.expression.toString())))
-      .toSet()
-      .toList();
-
-  /// Creates the parameters for the intl function call
-  /// ex: For the parameterized string 'Interpolated ${foo.bar} and $baz'
-  ///     it will return (foo.bar, baz)
-  String intlFunctionArguments(StringInterpolation node) {
-    var args = node.elements
-        .whereType<InterpolationExpression>()
-        .map((e) => e.expression.toString())
-        .toSet()
-        .toList();
-    return '(${args.join(', ')})';
-  }
-
-  String giveMeANameFor(StringLiteral body,
-      {String? initialName, bool startAtZero = false}) {
-    var messageText = body.toSource();
-    var functionName = toVariableName(messageText);
-    functionName = giveMeANameForString(
-        initialName ?? functionName, messageText,
-        startAtZero: startAtZero);
-    return functionName;
-  }
-
-  String giveMeANameForString(String name, String messageText,
+  /// Find the existing name for [name]+number that has the same [messageText], or
+  /// if there isn't one, return the first available.
+  String nameForString(String name, String messageText,
       {bool startAtZero = false}) {
     var index = 1;
     var newName = '$name${startAtZero ? 0 : ''}';
@@ -222,6 +88,7 @@ Attempting to add a different message with the same name:
     return newName;
   }
 
+  /// Is there already something with this name, but a different messageText?
   bool isNameTaken(String name, String messageText) {
     var method = methods[name];
     return method != null && methods[name]!.messageText != messageText;
@@ -269,6 +136,7 @@ class $className {''';
   String get prologue => prologueFor(className);
 }
 
+/// Holds the important parts of a method that we've parsed.
 class Method {
   String name;
   String messageText;
